+++ armadillo/src/main/java/at/favre/lib/armadillo/ByteArrayRuntimeObfuscator.java
@@ -1,6 +1,7 @@
 package at.favre.lib.armadillo;
 
 import java.security.SecureRandom;
+import java.util.Objects;
 
 import at.favre.lib.bytes.Bytes;
 
@@ -22,18 +23,60 @@ public interface ByteArrayRuntimeObfuscator {
      */
     byte[] getBytes();
 
-    final class Default implements ByteArrayRuntimeObfuscator {
+    /**
+     * Wipe and overwrite the internal byte arrays
+     */
+    void wipe();
 
+    final class Default implements ByteArrayRuntimeObfuscator {
         private final byte[][] data;
+        private final SecureRandom secureRandom;
 
         Default(byte[] array, SecureRandom secureRandom) {
-            byte[] key = Bytes.random(array.length, secureRandom).array();
-            this.data = new byte[][]{key, Bytes.wrap(array).mutable().xor(key).array()};
+            Objects.requireNonNull(array);
+            this.secureRandom = Objects.requireNonNull(secureRandom);
+            int keyCount = (int) (Math.abs(Bytes.random(8).toLong()) % 9) + 1;
+            this.data = new byte[keyCount + 1][];
+            createAndEncrypt(secureRandom, Bytes.from(array), array.length);
+        }
+
+        private void createAndEncrypt(SecureRandom secureRandom, Bytes from, int length) {
+            Bytes copy = from.mutable();
+            for (int i = 0; i < data.length - 1; i++) {
+                byte[] key = Bytes.random(length, secureRandom).array();
+                this.data[i] = key;
+                copy.xor(key);
+            }
+            this.data[data.length - 1] = copy.array();
         }
 
         @Override
         public byte[] getBytes() {
-            return Bytes.wrap(data[0]).xor(data[1]).array();
+            Bytes b = Bytes.empty();
+            for (int i = data.length - 1; i >= 0; i--) {
+                if (b.isEmpty()) {
+                    b = Bytes.from(data[i]).mutable();
+                    continue;
+                }
+                b.xor(data[i]);
+            }
+
+            createAndEncrypt(this.secureRandom, Bytes.from(b), b.length());
+
+            return b.array();
+        }
+
+        @Override
+        public void wipe() {
+            for (byte[] arr : data) {
+                Bytes.wrap(arr).mutable().secureWipe();
+            }
+        }
+
+        @Override
+        protected void finalize() throws Throwable {
+            wipe();
+            super.finalize();
         }
     }
 }