+++ armadillo/src/main/java/at/favre/lib/armadillo/EncryptionProtocol.java
@@ -87,6 +87,29 @@ interface EncryptionProtocol {
     KeyStretchingFunction getKeyStretchingFunction();
 
     /**
+     * Take raw char password and obfuscate (for use in-memory)
+     *
+     * @param password to obfuscate
+     * @return obfuscated password
+     */
+    @Nullable
+    ByteArrayRuntimeObfuscator obfuscatePassword(@Nullable char[] password);
+
+    /**
+     * Convert obfuscated obfuscated bytes to char array again
+     *
+     * @param obfuscated to deobfuscate
+     * @return char array
+     */
+    @Nullable
+    char[] deobfuscatePassword(@Nullable ByteArrayRuntimeObfuscator obfuscated);
+
+    /**
+     * If enabled, wipes the internal derived password cache.
+     */
+    void wipeDerivedPasswordCache();
+
+    /**
      * Factory creating a new instance of {@link EncryptionProtocol}
      */
     interface Factory {
@@ -119,5 +142,8 @@ interface EncryptionProtocol {
          * @return random gen
          */
         SecureRandom getSecureRandom();
+
+        @Nullable
+        ByteArrayRuntimeObfuscator obfuscatePassword(@Nullable char[] password);
     }
 }