+++ armadillo/src/main/java/at/favre/lib/armadillo/SecureSharedPreferences.java
@@ -41,7 +41,7 @@ public final class SecureSharedPreferences implements ArmadilloSharedPreferences
     private final RecoveryPolicy recoveryPolicy;
 
     @Nullable
-    private char[] password;
+    private ByteArrayRuntimeObfuscator password;
     private boolean supportVerifyPassword;
 
     private String prefSaltContentKey;
@@ -63,7 +63,7 @@ public final class SecureSharedPreferences implements ArmadilloSharedPreferences
         this.sharedPreferences = sharedPreferences;
         this.factory = encryptionProtocolFactory;
         this.recoveryPolicy = recoveryPolicy;
-        this.password = password;
+        this.password = factory.obfuscatePassword(password);
         this.supportVerifyPassword = supportVerifyPassword;
         init();
     }
@@ -295,9 +295,9 @@ public final class SecureSharedPreferences implements ArmadilloSharedPreferences
         }
 
         if (password != null) {
-            Arrays.fill(password, (char) 0);
+            password.wipe();
         }
-        password = newPassword;
+        password = encryptionProtocol.obfuscatePassword(newPassword);
     }
 
     /**
@@ -311,6 +311,7 @@ public final class SecureSharedPreferences implements ArmadilloSharedPreferences
      */
     private boolean reencryptStringType(@Nullable char[] newPassword, Editor editor, String keyHash, @Nullable KeyStretchingFunction newKsFunction) {
         try {
+            final ByteArrayRuntimeObfuscator pw = encryptionProtocol.obfuscatePassword(newPassword);
             final String encryptedValue = sharedPreferences.getString(keyHash, null);
 
             if (encryptedValue == null) {
@@ -326,7 +327,7 @@ public final class SecureSharedPreferences implements ArmadilloSharedPreferences
                 encryptionProtocol.setKeyStretchingFunction(newKsFunction);
             }
 
-            editor.putEncryptedBase64(keyHash, encryptToBase64(keyHash, newPassword, bytes));
+            editor.putEncryptedBase64(keyHash, encryptToBase64(keyHash, pw, bytes));
             return true;
         } catch (ClassCastException e) {
             return false;
@@ -343,6 +344,7 @@ public final class SecureSharedPreferences implements ArmadilloSharedPreferences
      * @return returns true if the StringSet was successfully re-encrypted.
      */
     private boolean reencryptStringSetType(char[] newPassword, Editor editor, String keyHash, @Nullable KeyStretchingFunction newKsFunction) {
+        final ByteArrayRuntimeObfuscator pw = encryptionProtocol.obfuscatePassword(newPassword);
         final Set<String> encryptedSet = sharedPreferences.getStringSet(keyHash, null);
         if (encryptedSet == null) {
             return false;
@@ -363,7 +365,7 @@ public final class SecureSharedPreferences implements ArmadilloSharedPreferences
 
         final Set<String> encryptedValues = new HashSet<>(decryptedSet.size());
         for (String value : decryptedSet) {
-            encryptedValues.add(encryptToBase64(keyHash, newPassword, Bytes.from(value).array()));
+            encryptedValues.add(encryptToBase64(keyHash, pw, Bytes.from(value).array()));
         }
         editor.putEncryptedStringSet(keyHash, encryptedValues);
         return true;
@@ -372,7 +374,7 @@ public final class SecureSharedPreferences implements ArmadilloSharedPreferences
     @Override
     public void close() {
         if (password != null) {
-            Arrays.fill(password, (char) 0);
+            password.wipe();
         }
         password = null;
         if (preferencesSalt != null) {
@@ -502,21 +504,22 @@ public final class SecureSharedPreferences implements ArmadilloSharedPreferences
     }
 
     @NonNull
-    private String encryptToBase64(String keyHash, @Nullable char[] password, byte[] content) {
+    private String encryptToBase64(String keyHash, @Nullable ByteArrayRuntimeObfuscator password, byte[] content) {
         try {
-            return Bytes.wrap(encryptionProtocol.encrypt(keyHash, password, content)).encodeBase64();
+            return Bytes.wrap(encryptionProtocol.encrypt(keyHash, encryptionProtocol.deobfuscatePassword(password), content)).encodeBase64();
         } catch (EncryptionProtocolException e) {
             throw new IllegalStateException(e);
         }
     }
 
     @Nullable
-    private byte[] decrypt(String keyHash, char[] password, @NonNull String base64Encrypted) {
+    private byte[] decrypt(String keyHash, @Nullable ByteArrayRuntimeObfuscator password, @NonNull String base64Encrypted) {
         try {
-            return encryptionProtocol.decrypt(keyHash, password, Bytes.parseBase64(base64Encrypted).array());
+            return encryptionProtocol.decrypt(keyHash, encryptionProtocol.deobfuscatePassword(password), Bytes.parseBase64(base64Encrypted).array());
         } catch (EncryptionProtocolException e) {
             recoveryPolicy.handleBrokenContent(e, keyHash, base64Encrypted, password != null, this);
         }
         return null;
     }
+
 }